+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.0.0 / 2020-12-2
+## 0.1.5 / 2020-12-2
 
 - proxy http & https
 - proxy lifecycle hooks
@@ -6,3 +6,8 @@
 - system web proxy control
 - import root certificate into system trust list
 - https certificate auto-generate
+
+## 0.1.6 / 2020-12-5
+
+- optimize the code, and remove @fxtop/winffi dependency
+- change the way to set windows web proxy, much better compatible